Rename the package and add more tests
