+++ package.json
@@ -1,6 +1,6 @@
 {
-  "name": "cache-manager-sqlite",
-  "version": "0.2.0",
+  "name": "@eligundry/cache-manager-better-sqlite",
+  "version": "1.0.0",
   "description": "A modern SQLite store for node-cache-manager",
   "main": "dist/index.js",
   "types": "./dist/index.d.ts",
