+++ tsconfig.build.json
@@ -1,6 +1,6 @@
 {
   "extends": "./tsconfig.json",
   "exclude": [
-    "./src/test/*"
+    "./src/tests/*"
   ]
 }
